Refactor of main
Refactor of the main.go source file.
