+++ go.mod
@@ -4,5 +4,4 @@ go 1.19
 
 require (
 	github.com/go-chi/chi/v5 v5.0.8 // indirect
-	github.com/gorilla/mux v1.8.0 // indirect
 )
