+++ src/testdata/attractionDetail_test.json
@@ -1,0 +1,54 @@
+[
+    {
+        "id": 1001,
+        "name": "Royal Observatory Greenwich",
+        "city": "London",
+        "country": "United Kingdom",
+        "hoursOpen": "09:00 - 18:00",
+        "daysOpen": "all",
+        "rating": 4.7,
+        "distance": 2.64,
+        "latitude": -0.000536,
+        "longitude": 51.476833,
+        "attractionImageName": "royalObservatory"
+    },
+    {
+        "id": 1002,
+        "name": "Cutty Sark",
+        "city": "London",
+        "country": "United Kingdom",
+        "hoursOpen": "09:00 - 17:00",
+        "daysOpen": "weekdays",
+        "rating": 4.5,
+        "distance": 1.85,
+        "latitude": -0.009586,
+        "longitude": 51.482880,
+        "attractionImageName": "cuttySark"
+    },
+    {
+        "id": 1003,
+        "name": "Tower Bridge",
+        "city": "London",
+        "country": "United Kingdom",
+        "hoursOpen": "09:00 - 17:30",
+        "daysOpen": "all",
+        "rating": 4.9,
+        "distance": 3.45,
+        "latitude": -0.075402,
+        "longitude": 51.505507,
+        "attractionImageName": "towerBridge"
+    },
+    {
+        "id": 1004,
+        "name": "O2 Millenium Dome",
+        "city": "London",
+        "country": "United Kingdom",
+        "hoursOpen": "09:00 - 18:00",
+        "daysOpen": "all",
+        "rating": 4.1,
+        "distance": 2.35,
+        "latitude": 0.003182,
+        "longitude": 51.502937,
+        "attractionImageName": "milleniumDome"
+    }
+]
