+++ src/testdata/attractionOutline_test.json
@@ -1,0 +1,42 @@
+[
+    {
+        "id": 1001,
+        "name": "Royal Observatory Greenwich",
+        "isOpen": true,
+        "rating": 4.7,
+        "distance": 2.64,
+        "latitude": -0.000536,
+        "longitude": 51.476833,
+        "attractionImageName": "royalObservatory"
+    },
+    {
+        "id": 1002,
+        "name": "Cutty Sark",
+        "isOpen": true,
+        "rating": 4.5,
+        "distance": 1.85,
+        "latitude": -0.009586,
+        "longitude": 51.482880,
+        "attractionImageName": "cuttySark"
+    },
+    {
+        "id": 1003,
+        "name": "Tower Bridge",
+        "isOpen": false,
+        "rating": 4.9,
+        "distance": 3.45,
+        "latitude": -0.075402,
+        "longitude": 51.505507,
+        "attractionImageName": "towerBridge"
+    },
+    {
+        "id": 1004,
+        "name": "O2 Millenium Dome",
+        "isOpen": true,
+        "rating": 4.1,
+        "distance": 2.35,
+        "latitude": 0.003182,
+        "longitude": 51.502937,
+        "attractionImageName": "milleniumDome"
+    }
+]
